requirements.txt
mne>=1.2
numpy
scipy
pandas
PyWavelets
matplotlib
pillow
torch
torchvision
tqdm
scikit-learn